+++ src/Exceptions/Auth/EmailTakenException.cs
@@ -1,0 +1,6 @@
+namespace social_media_backend.Exceptions.Auth;
+
+public class EmailTakenException : Exception
+{
+	
+}
