+++ src/Exceptions/Auth/InvalidCredentialsException.cs
@@ -1,4 +1,4 @@
-namespace social_media_backend.Exceptions;
+namespace social_media_backend.Exceptions.Auth;
 
 public class InvalidCredentialsException : Exception
 {
