+++ src/Exceptions/User/UserExistsException.cs
