+++ src/Services/AuthService.cs
@@ -1,76 +1,75 @@
 using MySql.Data.MySqlClient;
 using social_media_backend.Exceptions;
+using social_media_backend.Exceptions.Auth;
+using social_media_backend.src.Services;
 using social_media_backend.Util;
 
 namespace social_media_backend.Services;
 
 public class AuthService
 {
+	private readonly UserService _userService = new();
+	
 	public string Signup(string username, string name, string email, string password)
 	{
-		DatabaseService.OpenConnection();
-
-		if (DoesUserExist(username, email)) throw new UserExistsException();
-
-		int userId;
-		using (var command = new MySqlCommand("INSERT INTO users (username, name, email, password) VALUES (@username, @name, @email, @password); SELECT LAST_INSERT_ID();",
-			       DatabaseService.Connection))
+		try
 		{
-			command.Parameters.AddWithValue("@username", username);
-			command.Parameters.AddWithValue("@name", name);
-			command.Parameters.AddWithValue("@email", email);
-			command.Parameters.AddWithValue("@password", HashUtil.GenerateSHA256Hash(password));
+			DatabaseService.OpenConnection();
+			if (_userService.DoesUserExistByEmail(email)) throw new EmailTakenException();
+			if (_userService.DoesUserExistByUsername(username)) throw new UserExistsException();
+			
+			int userId;
+			using (var command = new MySqlCommand("INSERT INTO users (username, name, email, password) VALUES (@username, @name, @email, @password); SELECT LAST_INSERT_ID();",
+				       DatabaseService.Connection))
+			{
+				command.Parameters.AddWithValue("@username", username);
+				command.Parameters.AddWithValue("@name", name);
+				command.Parameters.AddWithValue("@email", email);
+				command.Parameters.AddWithValue("@password", HashUtil.GenerateSHA256Hash(password));
 
-			var result = command.ExecuteScalar();
-			userId = Convert.ToInt32(result); // Get the new user's ID
+				var result = command.ExecuteScalar();
+				userId = Convert.ToInt32(result); // Get the new user's ID
                 
+			}
+			return TokenUtil.CreateToken(userId.ToString(), email);
 		}
-
-		DatabaseService.CloseConnection();
-
-		return TokenUtil.CreateToken(userId.ToString(), email);
+		finally
+		{
+			DatabaseService.CloseConnection();
+		}
 	}
 
 	public string Login(string email, string password)
 	{
 		DatabaseService.OpenConnection();
-		
-		if(!DoesUserExist(email)) throw new InvalidCredentialsException();
-		
-		DatabaseService.OpenConnection();
-		var userId = -1;
-		var storedPasswordHash = string.Empty;
-		
-		using (var command = new MySqlCommand("SELECT id , email , password FROM users WHERE email = @email AND password = @password", DatabaseService.Connection))
-		{
-			command.Parameters.AddWithValue("@email", email);
-			command.Parameters.AddWithValue("@password", HashUtil.GenerateSHA256Hash(password));
-			var result = command.ExecuteReader();
-			if (result.Read())
-			{
-				userId = result.GetInt32("id");
-				storedPasswordHash = result.GetString("password");
-			}
-		}
-		DatabaseService.CloseConnection();
 
-		if (HashUtil.GenerateSHA256Hash(password) != storedPasswordHash) throw new InvalidCredentialsException();
+		try
+		{
+			if(!_userService.DoesUserExistByEmail(email)) throw new InvalidCredentialsException();
+		
+			var userId = -1;
+			var storedPasswordHash = string.Empty;
+		
+			using (var command = new MySqlCommand("SELECT id , email , password FROM users WHERE email = @email AND password = @password", DatabaseService.Connection))
+			{
+				command.Parameters.AddWithValue("@email", email);
+				command.Parameters.AddWithValue("@password", HashUtil.GenerateSHA256Hash(password));
+				var result = command.ExecuteReader();
+				if (result.Read())
+				{
+					userId = result.GetInt32("id");
+					storedPasswordHash = result.GetString("password");
+				}
+			}
 			
-		return TokenUtil.CreateToken(userId.ToString(), email);
-	}
-	private static bool DoesUserExist(string username, string email)
-	{
-		using var command = new MySqlCommand("SELECT COUNT(*) FROM users WHERE username = @username OR email = @email", DatabaseService.Connection);
-		command.Parameters.AddWithValue("@username", username);
-		command.Parameters.AddWithValue("@email", email);
-		var result = command.ExecuteScalar();
-		return Convert.ToInt32(result) > 0;
-	}
-	private static bool DoesUserExist(string email)
-	{
-		using var command = new MySqlCommand("SELECT COUNT(*) FROM users WHERE email = @email", DatabaseService.Connection);
-		command.Parameters.AddWithValue("@email", email);
-		var result = command.ExecuteScalar();
-		return Convert.ToInt32(result) > 0;
+			if (HashUtil.GenerateSHA256Hash(password) != storedPasswordHash) throw new InvalidCredentialsException();
+			
+			return TokenUtil.CreateToken(userId.ToString(), email);
+		}
+		finally
+		{
+			DatabaseService.CloseConnection();
+		}
+
 	}
 }
